requirements.freeze.txt
# These are the most recent releases of these packages as of 10/4/2019,
# *except* for some that are built into App Engine Standard Python 2.
# Those, we pin to App Engine's versions:
# https://cloud.google.com/appengine/docs/standard/python/tools/built-in-libraries-27
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
extras==1.0.0
fixtures==3.0.0
future==0.17.1
-e git+https://github.com/snarfed/gdata-python-client-1.git@1df4e1efea7e5cf2754bc7eec6c1ab48ab09e3b1#egg=gdata-master
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth==1.6.3
google-cloud-ndb==0.2.0
gunicorn==19.10.0
html5lib==1.0.1
httplib2==0.14.0
humanize==0.5.1
idna==2.8
Jinja2==2.10.1  # App Engine's; see above
linecache2==1.0.0
lxml==3.7.3
mf2py==1.1.2
mf2util==0.5.1
mox3==0.28.0
oauthlib==3.1.0
pbr==5.4.3
pyasn1-modules==0.2.6
pyasn1==0.4.7
PySocks==1.7.1
python-mimeparse==1.6.0
python-tumblpy==1.1.4
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.0
six==1.12.0  # App Engine's; see above
soupsieve==1.9.4
testtools==2.3.0
traceback2==1.4.0
tweepy==3.8.0
ujson==1.35  # App Engine's; see above
unittest2==1.1.0
uritemplate==3.0.0
urllib3==1.25.6
webapp2==3.0.0b1  # *not* App Engine's (2.5.2) for Python 3 compatibility
webencodings==0.5.1
WebOb==1.8.5  # *not* App Engine's (1.2.3) for more recent bug fixes
